backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Redis
redis>=5.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Rate limiting
slowapi>=0.1.9

# Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.1.0

# HTTP client (for proxying to chatbot/ops-console)
httpx>=0.26.0

# Email
aiosmtplib>=3.0.0

# Utils
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Report Automation Dependencies
python-pptx>=0.6.23
pandas>=2.0.0
matplotlib>=3.7.0
python-docx>=1.1.0
Pillow>=10.0.0
ollama>=0.1.0
requests>=2.31.0
openai>=1.0.0
